@@ -30,7 +30,9 @@ namespace Fido2Demo
             });
 
             // Use the in-memory implementation of IDistributedCache.
+            services.AddMemoryCache();
             services.AddDistributedMemoryCache();
+
             services.AddSession(options =>
             {
                 // Set a short timeout for easy testing.
@@ -52,7 +54,10 @@ namespace Fido2Demo
             })
             .AddCachedMetadataService(config =>
             {
-                config.AddFidoMetadataRepository();
+                config.AddFidoMetadataRepository(httpClientBuilder =>
+                {
+                    //TODO: any specific config you want for accessing the MDS
+                });
             });
         }
 
