@@ -6,24 +6,14 @@ namespace Fido2NetLib
 {
     internal class NullMetadataService : IMetadataService
     {
-        bool IMetadataService.ConformanceTesting()
+        public Task<MetadataBLOBPayloadEntry> GetEntryAsync(Guid aaguid, CancellationToken cancellationToken = default)
+        {
+            return Task.FromResult((MetadataBLOBPayloadEntry)null);
+        }
+
+        public bool ConformanceTesting()
         {
             return false;
-        }
-
-        MetadataBLOBPayloadEntry IMetadataService.GetEntry(Guid aaguid)
-        {
-            return null;
-        }
-
-        Task IMetadataService.InitializeAsync(CancellationToken cancellationToken)
-        {
-            return Task.CompletedTask;
-        }
-
-        bool IMetadataService.IsInitialized()
-        {
-            return true;
         }
     }
 }
